requirements.txt
transformers
torch
Pillow
python-dotenv
APScheduler
instagrapi
